// types/profile.interface.ts
import { Connection } from "./connection.interface";
import ReportRequest from "./report-request.interface";
import Report from "./report.interface";
import TradeRequests from "./trade-request.interface";
import { Trade } from "./trade.interface";

export default interface Profile {
	interestedIn: boolean | undefined;
	accountOwner: any;
	interestedInNote: string | undefined;
	reviewedBy: any;
	userId: any;
	isInterestInFunding: boolean | undefined;
	interestInFundingNote: string | undefined;
	interestedInFundingNote: string | undefined;
	isInterestedInFunding: boolean | undefined;
	businessStartDateNote: string | undefined;
	isBusinessStartDate: boolean | undefined;
	businessSizeNote: string | undefined;
	isBusinessSize: boolean | undefined;
	registeredStateNote: string | undefined;
	articlesOfIncorporationNote: string | undefined;
	isDocsReviewed: boolean | undefined;
	einDocumentsNote: any;
	reviewedByAdminName: string | undefined;
	id: string;
	factiivAddress: string;
	isFactiivAddress: boolean;
	businessName: string;
	isBusinessName: boolean;
	isOwner: boolean;
	isIsOwner: boolean;
	ownerName: string;
	isOwnerName: boolean;
	industry: string;
	isIndustry: boolean;
	street: string;
	isStreet: boolean;
	city: string;
	isCity: boolean;
	state: string;
	isState: boolean;
	zip: string;
	isZip: boolean;
	country: string;
	isCountry: boolean;
	phoneNumber: string;
	isPhoneNumber: boolean;
	website: string;
	isWebsite: boolean;
	email: string;
	isEmail: boolean;
	ein: string;
	isEin: boolean;
	einDocumentPath: string;
	einDocumentApproved: boolean;
	aoiDocumentPath: string;
	aoiDocumentApproved: boolean;
	imagePath: string;
	googleBusiness: string;
	isGoogleBusiness: boolean;
	phoneListing: string;
	isPhoneListing: boolean;
	lenderCounter: number;
	borrowerCounter: number;
	buyerCounter: number;
	sellerCounter: number;
	createdAt: string;
	reputationScore: number;
	factiivScore: number;
	utilizationScore: number;
	historyScore: number;
	verifiedStatus: boolean;
	profileDataStatus: boolean;
	profileDocsStatus: boolean;
	profileInfoStatus: boolean;
	isAnonymous: string;
	reviewedAt: string;
	submittedAt: string;
	docReviewedAt: string;
	docSubmittedAt: string;
	myTrades: Trade[];
	tradesWithMe: Trade[];
	reports: Report[];
	reportRequests: ReportRequest[];
	tradeRequests: TradeRequests[];
	connections: Connection[];
	admin: any;
	businessEINNote: string | undefined;
	businessWebsiteNote: string | undefined;
	businessEmailNote: string | undefined;
	businessPhoneNote: string | undefined;
	businessAddressNote: string | undefined;
	ownerNameNote: string | undefined;
	businessVerticalNote: string | undefined;
	isOwnerNote: string | undefined;
	businessNameNote: string | undefined;
	submitTime: string | undefined;
	businessVertical: string | undefined;
	businessAddress: string | undefined;
	businessPhone: string | undefined;
	businessEmail: string | undefined;
	businessWebsite: string | undefined;
	businessEIN: number | undefined;
	isReviewed: boolean | undefined;
	vendorAccounts: boolean;
	sbaLoans: boolean;
	linesOfCredit: boolean;
	businessSize: string;
	businessStartDate: string;
	interestedInFunding: boolean;
	businessCreditCards: boolean;
	interestInFunding: boolean;
	profileId: string;
	dataUnderReview: boolean;
};